+++ src/packages/api/account/services/get-updated-accounts.ts
@@ -29,7 +29,7 @@ export async function getUpdatedAccounts(
 	const response = await this.apiClient.runRequest(url, { method: 'GET' });
 
 	// Format the Response
-	const accounts = this.formatResponse(response);
+	const accounts = await this.formatResponse(response);
 
 	// Setup the Search Response
 	const searchResponse: SearchResponse<Account> = {