+++ src/packages/api/company-info/services/get-company-info.ts
@@ -23,7 +23,7 @@ export async function getCompanyInfo(this: CompanyInfoAPI, options: CompanyInfoO
 	const response = await this.apiClient.runRequest(url, { method: 'GET' });
 
 	// Format the Response
-	const companyInfo = this.formatResponse(response);
+	const companyInfo = await this.formatResponse(response);
 
 	// Return the Company Info
 	return companyInfo;