+++ src/packages/api/estimates/services/get-all-estimates.ts
@@ -21,7 +21,7 @@ export async function getAllEstimates(this: EstimateAPI, options: EstimateOption
 	const response = await this.apiClient.runRequest(url, { method: 'GET' });
 
 	// Format the Response
-	const estimates = this.formatResponse(response);
+	const estimates = await this.formatResponse(response);
 
 	// Setup the Search Response
 	const searchResponse: SearchResponse<Estimate> = {